further review fixes + adding more comments/inline docs

+++ packages/contracts/contracts/libs/v0.8.x/minter-libs/DALinLib.sol
@@ -1,8 +1,6 @@
 // SPDX-License-Identifier: LGPL-3.0-only
 // Created By: Art Blocks Inc.
 
-import "@openzeppelin-4.5/contracts/utils/math/SafeCast.sol";
-
 pragma solidity ^0.8.0;
 
 /**
@@ -73,8 +71,8 @@ library DALinLib {
         DAProjectConfig storage _DAProjectConfig
     ) internal view returns (uint256) {
         // move parameters to memory if used more than once
-        uint256 _timestampStart = uint256(_DAProjectConfig.timestampStart);
-        uint256 _timestampEnd = uint256(_DAProjectConfig.timestampEnd);
+        uint256 _timestampStart = _DAProjectConfig.timestampStart;
+        uint256 _timestampEnd = _DAProjectConfig.timestampEnd;
         uint256 _startPrice = _DAProjectConfig.startPrice;
         uint256 _basePrice = _DAProjectConfig.basePrice;
